refactor(test_dal): move fixtures to the conftest.py
